refactor(ts): migrar app root (main, App, routes, providers) a TypeScript
