+++ locuventas_frontend/index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root" ></div>
     <div id="portal-root"></div> <!-- <- Para pirulas de alertas -->
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
