--- locuventas_frontend/src/app/providers.jsx
@@ -1,10 +1,0 @@
-import { AuthProvider } from "@context/AuthContext";
-import { HeaderProvider } from "@context/HeaderContext";
-
-export function AppProviders({ children }) {
-  return (
-    <AuthProvider>
-      <HeaderProvider>{children}</HeaderProvider>
-    </AuthProvider>
-  );
-}
